EliteMobs 9.4.0: - [New] Readded The Colosseum, which has now been remade for EliteMobs 9 - [New] Added useModels option in CustomModelsConfig.yml - [New] Added the Swift Sneak enchantment from vanilla Minecraft as valid for loot - [New] Added a new kind of instanced entity for an upcoming new game mode - [New] Added a new field to store custom data into Elites, to be used via API for other plugins (used by the upcoming new game mode) - [New] Elite Script cooldowns, particle values, relative vectors, script targets and script zones now follow the elite scripting format of being able to be randomized - [New] Added cone shapes to elite zones for scripting - [New] Treasure chests can now use the new map-based loot system - [Fix] Fixed bug with browsing quests - [Fix] Fixed bug where events could theoretically happen in instanced zones - [Fix] Fixed bug with taze power - [Fix] Fixed bug with getting custom attribute modifiers for items from other plugins - [Fix] Fixed bug that would make chat tell people a quest can't be accepted yet when the player had a quest ready to be completed with that npc - [Fix] CustomModels now correctly has a .yml extension - [Fix] Curse of vanishing on old items now gets correctly applied to minecraft 1.21.4+ - [Fix] Fixed error spam with LAVA trail - [Fix] Fixed error spam due to a music bug

+++ src/main/java/com/magmaguy/elitemobs/utils/shapes/Cuboid.java
@@ -11,17 +11,17 @@ import java.util.List;
 import java.util.stream.Collectors;
 
 public class Cuboid extends Shape {
-    private final double x;
-    private final double y;
-    private final double xBorder;
-    private final double yBorder;
+    private final Float x;
+    private final Float y;
+    private final Float xBorder;
+    private final Float yBorder;
     private Location centerLocation = null;
     private List<Vector> locationVectors = null;
     private List<Vector> edgeVectors = null;
-    private double z;
-    private double zBorder;
+    private Float z;
+    private Float zBorder;
 
-    public Cuboid(Double x, Double y, Double z, Double xBorder, Double yBorder, Double zBorder, Location centerLocation) {
+    public Cuboid(Float x, Float y, Float z, Float xBorder, Float yBorder, Float zBorder, Location centerLocation) {
         this.x = x;
         this.y = y;
         this.z = z;
@@ -85,8 +85,8 @@ public class Cuboid extends Shape {
     private List<Vector> getLocationVectors() {
         if (locationVectors != null) return locationVectors;
         locationVectors = new ArrayList<>();
-        double xHalf = Math.floor(x / 2d);
-        double zHalf = Math.floor(z / 2d);
+        Float xHalf = (float) Math.floor(x / 2f);
+        Float zHalf = (float) Math.floor(z / 2f);
         for (int localX = 0; localX < x; localX++)
             for (int localZ = 0; localZ < z; localZ++)
                 for (int localY = 0; localY < y; localY++)
